Fix yet another bug in the SearchForm
cleaned_data['is_public'] is always None (if that field isn't present on
the form), so self.cleaned_data.get("is_public", ""), would never return
the default, the empty string. Since the value wasn't the empty string,
the conditional `is_public is not ""` would pass, which wasn't supposed
to happen.

For fun, I simplified the logic in the is_archived filtering while I was
messing around