Get user model via get_user_model in reports.models
First, it's generally better to use get_user_model rather than importing
the user model directly.

In this case, this also fixes a circular import issue. Note that the
underlying problem here is in elasticmodels, which actually *recommends*
circular imports as a *matter of course* rather than using a solution
that's not quite so brittle (and generally frowned upon).