Move reports.models.Report.generate_icon to reports.utils
The reason for doing this is to make it easier to create icons without
needing to create a `Report`. This also declutters the `Report` model,
which was a bit unwieldy before these icon-related changes were made.

Also, it's not clear that creating icons should be part of the model,
although, a `Report` requires its icon to be created in order to be
fully initialized.

There's some room for debate around this, but this implementation has
the advantage that we can more-easily play around with creating icons
from the shell. It also makes testing easier, since we can test icon
generation in isolation from the `Report` model.

- Add `reports.utils` module
- Add `generate_icon` function to `reports.utils`
- Make the implementation idependent of `Report`
- Clean up the implementation

OIH-39