Disable Haystack's real time signal processor in copy_records task
The previous approach of setting `HAYSTACK_SIGNAL_PROCESSOR = None`
wasn't actually doing anything because it was being set after the real
time signal processor was already installed.

The new approach adjusts the relevant settings *before* `django.setup()`
is called.

The biggest upside to this is that importing records is now much faster
(~30s instead of ~2.5m).