Only import new species from old site
Only add species that have not yet been imported fromt the old site;
don't also update species that have already been imported.

This change was made because Lindsey made a bunch of changes to species
on the beta site and we don't want to overwrite those. She indicated
that she would no longer edit species on the old site but that new
species could be created to go along with new reports.

Refs 2015-12-14 email