Rewrite home page view as DRF view
In the process, use `ReportSerializer` instead of `Report.to_json()` to
serialize `Report`s and remove `Report.to_json()`.

In addition, instead of picking 10 of the most recent 100 reports at
random, always show the latest 25 publicly-viewable reports. I'm not
sure why showing 10 reports at random would be desirable. Also, sorting
by PK is not the best way to get the latest reports. This seems to be
quite a bit faster than the previous version (probably because we're
only selecting 25 records instead of 100 (only to discard most of
them)).

This also removes manual caching of the 'reports/_popover.html'
template. We don't need to worry about doing that manually since we're
using the cached template loader in production.