Revamp how/when Report icons are generated
This replaces the caching approach implemented in 1cfc469.

Icons are now generated up front since we know what all the possible
icons are and there aren't that many of them.

To ensure icons stay fresh, there's a new generate_icons management
command that does just what it says on the tin. This command is
configured to run nightly at 3:31am, but it can also be run manually if
necessary. Regenerating icons nightly is somewhat analogous to
rebuilding the search index nightly.

Icons will also be generated in the init and copy_records tasks (at the
end of each).

Instead of (re)generating icons dynamically when Reports are initialized
*and* saved, only generate them on save. This is just being defensive,
but it shouldn't actually be necessary, since the set of all possible
icons is known up front.

Generated icon file names are no longer hashed because this doesn't
serve any real purpose. Instead, we combine the base name of category
icon with severity colors; these names are unique *and* readable.

The reasons for doing all this are for simplicity and performance.

The performance problems in the previous versions were mainly due to a
somewhat hidden "query in a loop" situation.