Add category dropdown & remove advanced search altogether
These are related because adding the category dropdown means the
advanced search isn't very useful any more. The advanced search allowed
for selecting specific species, but it was unwieldy.

In the new version, the idea is that results can be constrained to a
single category and then species can be found using a keyword search.

Some of the benefits of this change are:

- Code is simplified, easier to understand, and cleaner
- UI is easier for users to understand and use while still making it
  possible to get refined results
- It lets us remove a pile of CSS & JS
- The reports list/search page will now load even faster

If it turns out that the client needs to be able to select individual
species after all, we'll add a redesigned, user-friendly version.

As discussed with Lindsey W. via email dated 2016-01-26.