Make User.first_name and last_name fields nullable
This is to prevent an empty string from being stored in either the
`first_name` or `last_name` fields, which will list users with an empty
string in either field at the front of a user queryset.

Exposition: the issue here is that the User model is being overloaded
for both public users, whose "accounts" are created automatically when
they create reports, and registered staff/admin users. The former don't
have/need first and last names while the latter do.