Do user search only when `q` query parameter is present
That is, only query Elasticsearch for users when the `q` parameter is
present. If the other parameters are present without the `q` parameter,
query the database instead.

Basically, this just means that if the `Is manager` checkbox is checked
and there's no search term, we search the database instead of ES so the
managers list will be ordered by name (first, last) instead of by
relevance.