When auto-creating user, convert email address to lower case
This pertains to when a user is created automatically by submitting a
report or being invited to review a report.

Otherwise, it's too easy to inadvertantly create duplicate users.

NOTE: Technically, email addresses are case-sensitive, but I don't think
this is ever useful or relevant in practice.

XXX: There may be other places in the code where we need to make similar
changes.