Fix test for InviteForm/Invite.create()
I can't say I understand the use of the mock library in such cases. It
doesn't serve any purpose that I can see and, worse, seems to have a
tendency to hide bugs by not actually testing the code that needs to be
tested.

If this test had been working, the regression fixed in a2b6585 probably
wouldn't have happened.