Add SECURE_PROXY_SSL_HEADER stage setting
This is necessary because all of our stage sites run behind a proxy.

This fixes the issue with auto-login URLs not working. Previously, they
were being generated with http as the scheme instead of https. On
redirect from http to https, the URL would get reencoded.

I'm fairly certain this was a stage-only issue and we shouldn't need to
do anything in production.

OIH-49 #done