Disables uWSGI python auto reload.
Refs: https://github.com/PSU-OIT-ARC/oregon-registry-online/commit/5b7a642c8df7dd0e690a23c7c4cec8e651a4224c

It appears that there are strange interactions when running uWSGI with
the python-auto-reload option on Docker Engine 23.

When code changes are detected and the worker(s) are reloaded, traffic
between the proxy and application container appears to hang and the
processes representing the application containers continue to consume
100% of CPU resources.

Superficially, it appears this issue is related to seccomp changes
released in Docker Engine 23 and backported to Docker Engine 20 (though
subsequenctly reverted).

Refs: https://stackoverflow.com/questions/75511524/nginxuwsgi-stops-responding-after-auto-reload
Refs: https://docs.docker.com/engine/release-notes/23.0/
Refs: https://github.com/moby/moby/pull/44563
Refs: https://github.com/moby/moby/pull/44712