Adds new cloudformation defintions for 'vpc-cloud'.
The existing production environment resides in the original (legacy)
VPC 'vpc-prod' which is planned to be pruned of active projects.

Additionally, the EC2 instance is an AL2 host (which has since been
deprecated) and its RDS instance currently runs PostgreSQL 11 which
is EOL and will incur additional charges unless it is upgraded.