Uses conditional URL in 'notify_invited_reviewer'.
Both the 'notify_report_submission' and 'notify_report_subscribers'
tasks are implemented such that the included report URL is given
as a direct, absolute URL if the recipient is an existing staff member
and as a redirect for recipients who don't already have an account
in the system.

This latter functionality allows third-parties to
be securely logged-in for the purposes of collaboration on the
corresponding report. The URL used in this case contains cryptographic
information which is used to login the user and will expire after
a predetermined period of time.

In cases where existing staff are being invited to comment on reports,
it seems desireable to obviate the need for and the potential
frustration of expired links.

Reported-by: Jacob Rose <jacorose@pdx.edu>